requirements.txt
Keras==2.4.3
streamlit==0.76.0
numpy==1.18.5
tensorflow==2.3.1
opencv_python_headless==4.5.1.48
Pillow==8.1.2
